refactor: split android manager module
